+++ puncturedfem/plot/__init__.py
@@ -4,3 +4,11 @@ plot
 
 Subpackage for plotting traces, meshes, contour plots, heat maps, etc.
 """
+
+from .mesh_plot import MeshPlot
+from .trace_plot import TracePlot
+
+__all__ = [
+    "MeshPlot",
+    "TracePlot",
+]